Refactor directory name
