+++ import/options/FirefoxOptions.js
