+++ lib/restApi/ChromeApi.js
@@ -12,7 +12,7 @@
  @version   :   1.0
 */
 const {WebDriverRestApi} = require("./WebDriverRestApi.js");
-const {WebElement}       = require("../../socket/WebElement.js");
+const {WebElement}       = require("../../import/WebElement.js");
 
 const webDriverApi = WebDriverRestApi.getInstance();
